+++ src/Constructor/Editor.java
@@ -10,7 +10,7 @@ public class Editor {
 
     public static void menuEditor(Producto p) {
         Deposito.mostrarProducto(p);
-        int opcion = Validar.Int("\nQue desea editar: ");
+        byte opcion = Validar.Byte("\nQue desea editar: ");
         if(opcion >= 1 && opcion < 5){
             Basic(opcion, p);
         }
@@ -22,21 +22,21 @@ public class Editor {
         }
     }
 
-    public static void Carne(Carnes carne, int opcion){
+    public static void Carne(Carnes carne, byte opcion){
         switch (opcion){
             case 4 -> editarAtributo("Ingresa el nuevo porcentaje de Grasa: ", carne::setCantidadgrasa);
             case 5 -> editarAtributo("Ingresa la nueva cantidad de Proteínas: ", carne::setCantidadproteinas);
         }
     }
 
-    public static void Fruta(Fruta fruta, int opcion){
+    public static void Fruta(Fruta fruta, byte opcion){
         switch (opcion) {
             case 4 -> editarAtributo("Ingresa el nuevo porcentaje de Agua: ", fruta::setCantidadagua);
             case 5 -> editarAtributo("Ingresa la nueva cantidad de Azúcar: ", fruta::setCantidadazucar);
         }
     }
 
-    public static void Lacteo(Lacteo lacteo, int opcion) {
+    public static void Lacteo(Lacteo lacteo, byte opcion) {
         switch (opcion) {
             case 4 -> editarAtributo("Ingresa el nuevo porcentaje de Agua: ", lacteo::setPorcentajeagua);
             case 5 -> editarAtributo("Ingresa el nuevo porcentaje de Calcio: ", lacteo::setPorcentajecalcio);
@@ -44,7 +44,7 @@ public class Editor {
         }
     }
 
-    public static void Basic(int opcion, Producto p){
+    public static void Basic(byte opcion, Producto p){
         switch(opcion){
             case 1 -> p.setNombre(Validar.Nombre("\nIngresa el nuevo nombre del Producto: "));
             case 2 -> editarAtributo("Ingresa la nueva cantidad: ", p::setCantidad);
